[Feature] Create view extension functions for reusable styling of buttons
